refactor: Fake class 패키지 이동
