+++ app/src/test/java/com/sdhong/pokemonapp/viewmodel/HistoryViewModelTest.kt
@@ -1,6 +1,7 @@
 package com.sdhong.pokemonapp.viewmodel
 
 import com.sdhong.pokemonapp.local.model.Pokemon
+import com.sdhong.pokemonapp.viewmodel.fake.FakeHistoryDao
 import org.junit.jupiter.api.Assertions.assertEquals
 import org.junit.jupiter.api.Test
